+++ src/kagglehub/clients.py
@@ -74,11 +74,10 @@ def get_user_agent() -> str:
         str: user agent information.
     """
     user_agents = [f"kagglehub/{kagglehub.__version__}"]
-
-    for keras_lib in ("keras_hub", "keras_nlp", "keras_cv", "keras"):
-        keras_info = search_lib_in_call_stack(keras_lib)
-        if keras_info is not None:
-            user_agents.append(keras_info)
+    for lib in ("keras_hub", "keras_nlp", "keras_cv", "keras", "torchtune"):
+        lib_info = search_lib_in_call_stack(lib)
+        if lib_info is not None:
+            user_agents.append(lib_info)
             break
 
     # Add an appropriate data loader user agent for kagglehub.dataset_load calls